requirements.txt
# Core
PyYAML
openai>=1.14
python-dotenv
jsonschema

# Vector DB
faiss-cpu
sentence-transformers
numpy

# Sandbox
docker

# Async (Optional, for some tools)
aiofiles
aiohttp

# Web Scraping
requests
beautifulsoup4